Muestra funciones en data.js
