+++ src/main.js
@@ -1,4 +1,4 @@
-import { example } from "./data.js";
+import botonesPelis from "./data.js";
 import data from "./data/ghibli/ghibli.js";
 
 const peliculas = data.films;
@@ -6,8 +6,7 @@ const peliculas = data.films;
 const contenedorPeliculas = document.querySelector("#contenedor-peliculas");
 
 function renderPeliculas(peliculas) {
-  contenedorPeliculas.innerHTML = peliculas
-    .map((pelicula) => {
+  contenedorPeliculas.innerHTML = peliculas.map((pelicula) => {
       const { title, poster, director, release_date } = pelicula;
       return `<section class="grid-item">
         <p class="titulosPelis">${title}</p>
@@ -73,26 +72,11 @@ dropdowns.forEach((dropdown) => {
   });
 });
 
-// Función A-Z
 
-// Define la función "peliculasPorTituloAZ" que recibe el arreglo de objetos "peliculas" como parámetro.
-function peliculasPorTituloAZ(peliculas) {
-  // Utiliza el método "sort" del objeto "peliculas" para ordenar los elementos del arreglo según una función de comparación.
-  return peliculas.sort((a, b) => {
-    // Compara las propiedades "title" de los objetos "a" y "b" para determinar si "a.title" es menor que "b.title".
-    if (a.title < b.title) {
-      // Si "a.title" es menor que "b.title", retorna -1 para que "a" se ordene antes que "b".
-      return -1;
-    }
-    // Compara las propiedades "title" de los objetos "a" y "b" para determinar si "a.title" es mayor que "b.title".
-    if (a.title > b.title) {
-      // Si "a.title" es mayor que "b.title", retorna 1 para que "a" se ordene después que "b".
-      return 1;
-    }
-    // Si "a.title" es igual a "b.title", retorna 0 para mantener el orden actual.
-    return 0;
-  });
-}
+
+
+
+
 
 // Botón A-Z
 
@@ -102,26 +86,16 @@ const botonAZ = document.querySelector("li:nth-child(3)");
 // Agregar un evento "click" al botón "A-Z"
 botonAZ.addEventListener("click", () => {
   // Llamar a la función "peliculasPorTituloAZ" para ordenar las películas por orden alfabético (de A a Z)
-  const peliculasAZ = peliculasPorTituloAZ(peliculas);
+  const peliculasAZ = botonesPelis.peliculasPorTituloAZ(peliculas);
   // Renderizar las películas ordenadas por título
   renderPeliculas(peliculasAZ);
 });
 
-// Función Z-A
 
-// Definir la función "peliculasPorTituloZA" para ordenar las películas por orden alfabético inverso (de Z a A)
-function peliculasPorTituloZA(peliculas) {
-  // Usar el método "sort" para ordenar las películas por título
-  return peliculas.sort((a, b) => {
-    if (a.title < b.title) {
-      return 1; // Si el título de la película "a" es menor que el título de la película "b", devuelve 1
-    }
-    if (a.title > b.title) {
-      return -1; // Si el título de la película "a" es mayor que el título de la película "b", devuelve -1
-    }
-    return 0; // Si los títulos son iguales, devuelve 0
-  });
-}
+
+
+
+
 
 // Botón Z-A
 
@@ -131,22 +105,45 @@ const botonZA = document.querySelector("li:nth-child(4)");
 // Agregar un evento "click" al botón "Z-A"
 botonZA.addEventListener("click", () => {
   // Llamar a la función "peliculasPorTituloZA" para ordenar las películas por orden alfabético inverso (de Z a A)
-  const peliculasZA = peliculasPorTituloZA(peliculas);
+const peliculasZA = botonesPelis.peliculasPorTituloZA(peliculas);
 
   // Renderizar las películas ordenadas por título
   renderPeliculas(peliculasZA);
 });
 
-// Filtro directores
 
-// Usar el método "reduce" para crear un objeto con cada director de las películas
-const directores = peliculas.reduce((obj, pelicula) => {
-  obj[pelicula.director] = true;
-  return obj;
-}, {});
 
-// Obtener la lista de directores del objeto creado anteriormente
-const listaDirectores = Object.keys(directores);
+
+
+
+
+// Botón año ascendente
+const fechaOrdenarA = document.querySelector("li:nth-child(1)");
+
+fechaOrdenarA.addEventListener("click", () => {
+  const peliculasAño = botonesPelis.peliculasAscendente(peliculas);
+  renderPeliculas(peliculasAño);
+
+});
+
+
+
+
+
+
+
+// Botón año descendente
+const fechaOrdenarD = document.querySelector("li:nth-child(2)");
+
+fechaOrdenarD.addEventListener("click", () => {
+  const peliculasAñoD = botonesPelis.peliculasDescendente(peliculas);
+  renderPeliculas(peliculasAñoD);
+
+});
+
+
+
+// Botón directores
 
 // Obtener el elemento del botón desplegable por id
 const directorDropdown = document.querySelector("#directorDropdown");
@@ -170,67 +167,6 @@ directorOptions.forEach((option) => {
 
 
 
-// Función ordenar por año, orden Ascendente >
-
-function peliculasAscendente(peliculas) {
-  
-  return peliculas.sort((a, b) => {
-
-    const fechaA = new Date(a.release_date)
-    const fechaB = new Date(b.release_date)
-
-    if (fechaA < fechaB) {
-      return -1;
-    } else if (fechaA > fechaB) {
-      return 1;
-    } else {
-      return 0; 
-    }
-  });
-}
 
 
-// Botón año ascendente
-const fechaOrdenarA = document.querySelector("li:nth-child(1)");
-
-fechaOrdenarA.addEventListener("click", () => {
-  const peliculasAño = peliculasAscendente(peliculas);
-  renderPeliculas(peliculasAño);
-
-});
-
-
-
-
-// Función ordenar por año, orden descendente >
-
-function peliculasDescendente(peliculas) {
-  
-  return peliculas.sort((a, b) => {
-
-    const fechaA = new Date(a.release_date)
-    const fechaB = new Date(b.release_date)
-
-    if (fechaA < fechaB) {
-      return 1;
-    } if (fechaA > fechaB) {
-      return -1;
-    } 
-      return 0; 
-  });
-}
-
-
-// Botón año descendente
-const fechaOrdenarD = document.querySelector("li:nth-child(2)");
-
-fechaOrdenarD.addEventListener("click", () => {
-  const peliculasAñoD = peliculasDescendente(peliculas);
-  renderPeliculas(peliculasAñoD);
-
-});
-
-
-
-
-
+export default renderPeliculas;
