+++ src/style.css
@@ -85,7 +85,7 @@ footer {
 
 .anio {
   position: absolute;
-  top: 60%;
+  top: 70%;
   left: 50%;
   transform: translate(-50%, -50%);
 }
